finish 04-exercícios

+++ Aula05/atividades/04-exercícios/Q03.py
@@ -1,8 +1,6 @@
 # As maçãs custam 1,30 cada, se forem compradas menos de uma dúzia, e 1,00 se forem compradas pelo menos 12. Escreva um programa que leia o número de maçãs compradas, calcule e escreva o custo total da compra.
 
-
-
-maçãs = float(input("Quantas maçãs você quer comprar: "))
+maçãs = int(input("Quantas maçãs você quer comprar: "))
 
 if maçãs < 12:
     valor_maça = maçãs * 1.30
